+++ backend/src/db/base_model.py
@@ -1,0 +1,94 @@
+"""Base models for use in the other modules."""
+
+from __future__ import annotations
+
+from datetime import timedelta
+from math import floor
+from typing import TYPE_CHECKING, Any, override
+from uuid import UUID, uuid4
+
+from sqlalchemy.ext.asyncio import AsyncSession, async_object_session
+from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
+from sqlalchemy.types import Integer, TypeDecorator, TypeEngine
+
+if TYPE_CHECKING:
+    from sqlalchemy import Dialect
+
+
+class _TimedeltaAsMilliseconds(TypeDecorator[Integer]):
+    """Converts integer number of milliseconds to a Python timedelta object.
+
+    This class is used for converting values to and from the model database.
+    """
+
+    impl: TypeEngine[Any] | type[TypeEngine[Any]] = Integer
+    cache_ok: bool | None = True
+
+    @override
+    def process_bind_param(self, value: Any | None, dialect: Dialect) -> Any:
+        if value is not None:
+            if not isinstance(value, timedelta):
+                raise TypeError()
+            return floor(value.total_seconds() * 1000)
+        return value
+
+    @override
+    def process_result_value(self, value: Any | None, dialect: Dialect) -> Any | None:
+        if not isinstance(value, (float, int)):
+            raise TypeError()
+        return timedelta(milliseconds=value)
+
+
+class BaseSQLModel(DeclarativeBase):
+    """The base model for all models in the database.
+
+    This model has a standard SQL `id` field. It also includes a type annotation map to
+    convert Python timedelta objects to an integer number of milliseconds.
+
+    """
+
+    uuid: Mapped[UUID] = mapped_column(default=uuid4, primary_key=True)
+
+    __abstract__: bool = True
+    __type_annotation_map__: dict = {timedelta: _TimedeltaAsMilliseconds}
+
+    def get_parents(self) -> tuple[BaseSQLModel, ...]:
+        """Asynchronously get a tuple of this model's direct parents.
+
+        Returns:
+            tuple[BaseSQLModel]: the immediate parents of this model.
+
+        """
+        raise NotImplementedError('get_parents() is not implemented in this model')
+
+    def get_recursive_parents(self) -> tuple[BaseSQLModel, ...]:
+        """Recursively and asynchronously get a tuple of this model's parents.
+
+        This method is used to get the hierarchical branch of models that this model
+        is on. This is useful to ensure that clients can refresh objects that have
+        updated.
+
+        Returns:
+            tuple[BaseSQLModel]: the recursive parents of this model.
+
+        """
+        recursive_parents: list[BaseSQLModel] = list(self.get_parents())
+        for parent in self.get_parents():
+            if isinstance(parent, BaseSQLModel):
+                recursive_parents.extend(parent.get_recursive_parents())
+        return tuple(recursive_parents)
+
+    def get_session(self) -> AsyncSession:
+        """Get the session from the SQLAlchemy object.
+
+        Raises:
+            TypeError: if this object is not associated with a SQLAlchemy session.
+
+        Returns:
+            AsyncSession: the session with which this object is associated.
+
+        """
+        session: AsyncSession | None = async_object_session(self)
+        if session is None:
+            raise TypeError('This object is not associated with a SQL session')
+        return session
