+++ backend/src/db/engine.py
@@ -4,93 +4,23 @@ from __future__ import annotations
 
 import logging
 import os
-from datetime import timedelta
-from typing import (
-    TYPE_CHECKING,
-    ClassVar,
-    Final,
-)
-from uuid import UUID, uuid4
+from pathlib import Path
+from typing import TYPE_CHECKING, ClassVar, Final
 
 from sqlalchemy.engine import URL
-from sqlalchemy.ext.asyncio import (
-    AsyncSession,
-    async_object_session,
-    async_sessionmaker,
-    create_async_engine,
-)
-from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
+from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker, create_async_engine
 
 if TYPE_CHECKING:
     from pathlib import Path
 
     from sqlalchemy.ext.asyncio import AsyncEngine
+    from sqlalchemy.orm import DeclarativeBase
 
-
-from .utils import _TimedeltaAsMilliseconds
-
-CASCADE_CHILD: Final[str] = 'all, delete-orphan'
-CASCADE_OTHER: Final[str] = 'expunge, save-update'
 
 SQLALCHEMY_DEBUG: bool = os.environ.get('SQLALCHEMY_DEBUG', '').lower() in {
     'true',
     'yes',
 }
-
-
-class BaseSQLModel(DeclarativeBase):
-    """The base model for all models in the database.
-
-    This model has a standard SQL `id` field. It also includes a type annotation map to
-    convert Python timedelta objects to an integer number of milliseconds.
-
-    """
-
-    uuid: Mapped[UUID] = mapped_column(default=uuid4, primary_key=True)
-
-    __abstract__: bool = True
-    __type_annotation_map__: dict = {timedelta: _TimedeltaAsMilliseconds}
-
-    def get_parents(self) -> tuple[BaseSQLModel, ...]:
-        """Asynchronously get a tuple of this model's direct parents.
-
-        Returns:
-            tuple[BaseSQLModel]: the immediate parents of this model.
-
-        """
-        raise NotImplementedError('get_parents() is not implemented in this model')
-
-    def get_recursive_parents(self) -> tuple[BaseSQLModel, ...]:
-        """Recursively and asynchronously get a tuple of this model's parents.
-
-        This method is used to get the hierarchical branch of models that this model
-        is on. This is useful to ensure that clients can refresh objects that have
-        updated.
-
-        Returns:
-            tuple[BaseSQLModel]: the recursive parents of this model.
-
-        """
-        recursive_parents: list[BaseSQLModel] = list(self.get_parents())
-        for parent in self.get_parents():
-            if isinstance(parent, BaseSQLModel):
-                recursive_parents.extend(parent.get_recursive_parents())
-        return tuple(recursive_parents)
-
-    def get_session(self) -> AsyncSession:
-        """Get the session from the SQLAlchemy object.
-
-        Raises:
-            TypeError: if this object is not associated with a SQLAlchemy session.
-
-        Returns:
-            AsyncSession: the session with which this object is associated.
-
-        """
-        session: AsyncSession | None = async_object_session(self)
-        if session is None:
-            raise TypeError('This object is not associated with a SQL session')
-        return session
 
 
 class DatabaseEngine:
