+++ backend/src/game/bouts/models.py
@@ -6,7 +6,7 @@ from datetime import datetime  # noqa: TC003
 from typing import TYPE_CHECKING, Any, ClassVar, Final, final, override
 from uuid import UUID  # noqa: TC003
 
-from core import CASCADE_CHILD, CASCADE_OTHER, CacheableSQLModel, CacheKey
+from db import CASCADE_CHILD, CASCADE_OTHER, CacheableSQLModel
 from game.clocks.models import Clock
 from sqlalchemy import ForeignKey, column
 from sqlalchemy.orm import Mapped, mapped_column, relationship
@@ -15,7 +15,8 @@ from .schemas import BoutSchema
 from .types import BoutStateStr  # noqa: TC001
 
 if TYPE_CHECKING:
-    from core import BaseSQLModel
+    from core import CacheKey
+    from db import BaseSQLModel
     from game.jams.models import BaseJam
     from game.rulesets.schemas import Ruleset
     from game.series.models import Series
