+++ backend/src/game/series/models.py
@@ -4,12 +4,13 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, override
 
-from core import CASCADE_CHILD, BaseSQLModel, CacheableSQLModel, CacheKey
+from db import CASCADE_CHILD, BaseSQLModel, CacheableSQLModel
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from .schemas import SeriesSchema
 
 if TYPE_CHECKING:
+    from core import CacheKey
     from game.bouts.models import BaseBout
